requirements.txt
requests
google-generativeai
pandas
uuid
fastapi
uvicorn
pydantic
pyyaml
python-dotenv
python-multipart
pytest
httpx
pytest-asyncio
oracledb